:recycle: Refactor: 개인 관련 API 위치 이동
<body>
- :external-api:/api/individual

<footer>
- 관련: #141
