+++ application/external-api/src/main/java/com/namo/spring/application/external/api/individual/controller/PaletteController.java
@@ -1,0 +1,4 @@
+package com.namo.spring.application.external.api.individual.controller;
+
+public class PaletteController {
+}
