+++ application/external-api/src/main/java/com/namo/spring/application/external/api/individual/converter/PaletteConverter.java
@@ -1,0 +1,4 @@
+package com.namo.spring.application.external.api.individual.converter;
+
+public class PaletteConverter {
+}
