+++ application/external-api/src/main/java/com/namo/spring/application/external/api/individual/converter/ScheduleConverter.java
@@ -1,9 +1,9 @@
-package com.namo.spring.application.external.domain.individual.application.converter;
+package com.namo.spring.application.external.api.individual.converter;
 
-import com.namo.spring.application.external.domain.individual.domain.Schedule;
-import com.namo.spring.application.external.domain.individual.domain.constant.Period;
-import com.namo.spring.application.external.domain.individual.ui.dto.ScheduleRequest;
-import com.namo.spring.application.external.domain.individual.domain.Category;
+import com.namo.spring.db.mysql.domains.individual.domain.Schedule;
+import com.namo.spring.db.mysql.domains.individual.type.Period;
+import com.namo.spring.application.external.api.individual.dto.ScheduleRequest;
+import com.namo.spring.db.mysql.domains.individual.domain.Category;
 
 import com.namo.spring.db.mysql.domains.user.domain.User;
