+++ application/external-api/src/main/java/com/namo/spring/application/external/api/individual/dto/CategoryResponse.java
@@ -1,4 +1,4 @@
-package com.namo.spring.application.external.domain.individual.ui.dto;
+package com.namo.spring.application.external.api.individual.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
