+++ application/external-api/src/main/java/com/namo/spring/application/external/api/individual/facade/ScheduleFacade.java
@@ -1,4 +1,4 @@
-package com.namo.spring.application.external.domain.individual.application;
+package com.namo.spring.application.external.api.individual.facade;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -12,21 +12,21 @@ import com.namo.spring.application.external.domain.group.application.impl.MoimSc
 import com.namo.spring.application.external.domain.group.application.impl.MoimScheduleService;
 import com.namo.spring.application.external.domain.group.domain.MoimSchedule;
 import com.namo.spring.application.external.domain.group.domain.MoimScheduleAndUser;
-import com.namo.spring.application.external.domain.individual.application.converter.AlarmConverter;
-import com.namo.spring.application.external.domain.individual.application.converter.ScheduleConverter;
-import com.namo.spring.application.external.domain.individual.application.converter.ScheduleResponseConverter;
-import com.namo.spring.application.external.domain.individual.application.impl.AlarmService;
-import com.namo.spring.application.external.domain.individual.application.impl.CategoryService;
-import com.namo.spring.application.external.domain.individual.application.impl.ImageService;
-import com.namo.spring.application.external.domain.individual.application.impl.PeriodService;
-import com.namo.spring.application.external.domain.individual.application.impl.ScheduleService;
-import com.namo.spring.application.external.domain.individual.domain.Alarm;
-import com.namo.spring.application.external.domain.individual.domain.Category;
-import com.namo.spring.application.external.domain.individual.domain.Image;
-import com.namo.spring.application.external.domain.individual.domain.Schedule;
-import com.namo.spring.application.external.domain.individual.domain.constant.Period;
-import com.namo.spring.application.external.domain.individual.ui.dto.ScheduleRequest;
-import com.namo.spring.application.external.domain.individual.ui.dto.ScheduleResponse;
+import com.namo.spring.application.external.api.individual.converter.AlarmConverter;
+import com.namo.spring.application.external.api.individual.converter.ScheduleConverter;
+import com.namo.spring.application.external.api.individual.converter.ScheduleResponseConverter;
+import com.namo.spring.application.external.api.individual.service.AlarmService;
+import com.namo.spring.application.external.api.individual.service.CategoryService;
+import com.namo.spring.application.external.api.individual.service.ImageService;
+import com.namo.spring.application.external.api.individual.service.PeriodService;
+import com.namo.spring.application.external.api.individual.service.ScheduleService;
+import com.namo.spring.db.mysql.domains.individual.domain.Alarm;
+import com.namo.spring.db.mysql.domains.individual.domain.Category;
+import com.namo.spring.db.mysql.domains.individual.domain.Image;
+import com.namo.spring.db.mysql.domains.individual.domain.Schedule;
+import com.namo.spring.db.mysql.domains.individual.type.Period;
+import com.namo.spring.application.external.api.individual.dto.ScheduleRequest;
+import com.namo.spring.application.external.api.individual.dto.ScheduleResponse;
 import com.namo.spring.application.external.api.user.service.UserService;
 import com.namo.spring.db.mysql.domains.user.domain.User;
 import com.namo.spring.application.external.global.common.constant.FilePath;
