+++ application/external-api/src/main/java/com/namo/spring/application/external/api/individual/service/PeriodService.java
@@ -1,8 +1,8 @@
-package com.namo.spring.application.external.domain.individual.application.impl;
+package com.namo.spring.application.external.api.individual.service;
 
 import org.springframework.stereotype.Service;
 
-import com.namo.spring.application.external.domain.individual.domain.constant.Period;
+import com.namo.spring.db.mysql.domains.individual.type.Period;
 import com.namo.spring.core.common.code.status.ErrorStatus;
 import com.namo.spring.core.common.exception.IndividualException;
