--- application/external-api/src/main/java/com/namo/spring/application/external/domain/individual/application/converter/PaletteConverter.java
@@ -1,4 +1,0 @@
-package com.namo.spring.application.external.domain.individual.application.converter;
-
-public class PaletteConverter {
-}
