--- application/external-api/src/main/java/com/namo/spring/application/external/domain/individual/ui/PaletteController.java
@@ -1,4 +1,0 @@
-package com.namo.spring.application.external.domain.individual.ui;
-
-public class PaletteController {
-}
